requirements.txt
Flask==2.0.2
mysql-connector-python
ddtrace
